remove duplicated utils file
